requirements.txt
customtkinter>=5.2.0
pyserial>=3.5